% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sparsevinereg.R
\name{predict.sparsevinereg}
\alias{predict.sparsevinereg}
\title{Ppredictions with the vine copula based regression model}
\usage{
\method{predict}{sparsevinereg}(object, data, alpha, ...)
}
\arguments{
\item{object}{an object of class \code{sparsevinereg}.}

\item{data}{matrix or data frame of covariate values to use in predictions.}

\item{alpha}{vector of (quantile) prediction levels. \code{NA} predicts the mean
based on an average of the \code{1:10 / 11}-quantiles.}

\item{...}{unused.}
}
\value{
A data frame of predictions, where a column corresponds to a given
quantile level of \code{alpha}.
}
\description{
Ppredictions with the vine copula based regression model
}
\examples{
# Simulate data with 5 relevant and 5 irrelevant variables
set.seed(11)
x <- matrix(rnorm(5000), 500, 10)
y <- x[,1] -2*x[,2] + 3*x[,3] + 5*x[,4] - 4*x[,5]

 # response needs to be in the first column of the data
  data <- data.frame(y = y, x = x)

 # fit a sparse vine copula based regression model
 fit <- sparsevinereg(data)

 # print the final model
 print(fit)

 # make predictions with the model
 median_pred <- predict(fit, data = data, alpha = 0.5) # median
 mean_pred <- predict(fit, data = data, alpha = NA) # mean

}
